move condition related helpers to one file
